+++ packages/signals/lib/src/async/async_signal_state.dart
@@ -1,6 +1,15 @@
+/// Error builder for [AsyncSignalState]
 typedef AsyncSignalErrorBuilder<E> = E Function(
-    Object error, StackTrace? stackTrace);
-typedef AsyncSignalValueBuilder<E, T> = E Function(T value);
+  Object error,
+  StackTrace? stackTrace,
+);
+
+/// Value builder for [AsyncSignalState]
+typedef AsyncSignalValueBuilder<E, T> = E Function(
+  T value,
+);
+
+/// Generic builder for [AsyncSignalState]
 typedef AsyncSignalBuilder<E> = E Function();
 
 /// State for an [AsyncSignal]
@@ -16,38 +25,63 @@ sealed class AsyncSignalState<T> {
   })  : _value = value,
         _error = error;
 
+  /// Create a state with a value
   factory AsyncSignalState.data(T data) {
     return AsyncSignalStateData<T>(data);
   }
 
+  /// Create a state with an error
   factory AsyncSignalState.error(Object error, StackTrace? stackTrace) {
     return AsyncSignalStateError<T>(error, stackTrace);
   }
 
+  /// Create a state with a loading state
   factory AsyncSignalState.loading() {
     return AsyncSignalStateLoading<T>();
   }
 
+  /// Returns true if the state has a value
   bool get hasValue;
 
+  /// Returns true if the state has an error
   bool get hasError;
 
+  /// Returns true if the state is refreshing with a loading flag,
+  ///  has a value or error and is not the loading state
   bool get isRefreshing =>
       isLoading &&
       (hasValue || hasError) &&
       this is! AsyncSignalStateLoading<T>;
 
+  /// Returns true if the state is reloading with having a value or error,
+  /// and is the loading state
   bool get isReloading =>
       (hasValue || hasError) && this is AsyncSignalStateLoading<T>;
 
+  /// Force unwrap the value of the state.
+  ///
+  /// This will throw an error if the state does not have a value.
   T get requireValue => _value!;
 
+  /// Returns the value of the state.
   T? get value => _value;
 
+  /// Returns the error of the state.
   Object? get error => _error?.$1;
 
+  /// Returns the stack trace of the state.
   StackTrace? get stackTrace => _error?.$2;
 
+  /// Map the state to a value.
+  ///
+  /// ```dart
+  /// final signal = StreamSignal<int>();
+  /// signal.value.map(
+  ///  data: (value) => 'Value: $value',
+  ///  error: (error, stackTrace) => 'Error: $error',
+  ///  loading: () => 'Loading...',
+  /// );
+  /// ```
   E map<E>({
     required AsyncSignalValueBuilder<E, T> data,
     required AsyncSignalErrorBuilder<E> error,
@@ -62,6 +96,15 @@ sealed class AsyncSignalState<T> {
     return loading();
   }
 
+  /// Map the state to a value with optional or else.
+  ///
+  /// ```dart
+  /// final signal = StreamSignal<int>();
+  /// signal.value.maybeMap(
+  ///  data: (value) => 'Value: $value',
+  ///  orElse: () => 'Loading...',
+  /// );
+  /// ```
   E maybeMap<E>({
     AsyncSignalValueBuilder<E, T>? data,
     AsyncSignalErrorBuilder<E>? error,
@@ -105,19 +148,13 @@ sealed class AsyncSignalState<T> {
   }
 }
 
+/// State for an [AsyncSignal] with a value
 class AsyncSignalStateData<T> extends AsyncSignalState<T> {
   AsyncSignalStateData(
     T data, {
     super.isLoading = false,
     super.error,
   }) : super(value: data);
-
-  AsyncSignalStateData<T> toLoading() {
-    return AsyncSignalStateData<T>(
-      requireValue,
-      isLoading: true,
-    );
-  }
 
   @override
   bool get hasValue => true;
@@ -126,6 +163,7 @@ class AsyncSignalStateData<T> extends AsyncSignalState<T> {
   bool get hasError => false;
 }
 
+/// State for an [AsyncSignal] with an error
 class AsyncSignalStateError<T> extends AsyncSignalState<T> {
   AsyncSignalStateError(
     Object error,
@@ -134,14 +172,6 @@ class AsyncSignalStateError<T> extends AsyncSignalState<T> {
     super.value,
   }) : super(error: (error, stackTrace));
 
-  AsyncSignalStateError<T> toLoading() {
-    return AsyncSignalStateError<T>(
-      error!,
-      stackTrace,
-      isLoading: true,
-    );
-  }
-
   @override
   bool get hasValue => false;
 
@@ -149,16 +179,19 @@ class AsyncSignalStateError<T> extends AsyncSignalState<T> {
   bool get hasError => true;
 }
 
+/// State for an [AsyncSignal] with a loading state
 class AsyncSignalStateLoading<T> extends AsyncSignalState<T> {
   AsyncSignalStateLoading({
     super.value,
     super.error,
     super.isLoading = true,
+    this.hasValue = false,
+    this.hasError = false,
   });
 
   @override
-  bool get hasValue => false;
+  final bool hasValue;
 
   @override
-  bool get hasError => false;
+  final bool hasError;
 }